Added random loot in science lab.
Added some random loot in the chest in the science lab. At the moment
the code is quite long but I will be simplifying it next update, as well
as making more structures easier in the future. Also the items are
currently only spawning in the top half of the chest at the moment due
to a weird glitch when I try to add stuff in the other chest but
hopefully that will be fixed next time I work on this too. That's all.